requirements.txt
asgiref==3.10.0
Django==5.2.7
PyMySQL==1.1.0
cryptography==41.0.7
python-decouple==3.8
sqlparse==0.5.3
django-cors-headers==4.3.1
djangorestframework==3.15.2
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
dj-database-url==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0
setuptools>=65.0.0
requests==2.32.3
celery==5.4.0
redis==5.0.1
django-redis==5.4.0
